refactor: move __test__ folder inside src and modify ts-config root folder to server/src
