+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "node server/dist/src/index.js",
+    "start": "node server/dist/index.js",
     "dev": "cross-env NODE_ENV=dev ts-node-dev server/src/index.ts",
     "build": "tsc -p .",
     "react": "cd client && npm start",
@@ -14,7 +14,7 @@
     "lint:client": "eslint --ext .jsx client/src/",
     "lint-fix:client": "eslint --ext .jsx client/src/ --fix",
     "lint-fix": "eslint --ext .ts server/src/ --fix",
-    "test": "cross-env NODE_ENV=test jest server/__test__/",
+    "test": "cross-env NODE_ENV=test jest --runInBand server/src/__test__/",
     "prepare": "is-ci || npx husky install"
   },
   "repository": {
