+++ server/src/__test__/errors.test.ts
@@ -1,7 +1,7 @@
 /* eslint-disable no-undef */
 // eslint-disable-next-line import/no-extraneous-dependencies
 import supertest from 'supertest';
-import app from '../src/app';
+import app from '../app';
 
 describe('Get /gdfgjdfgkdfgkdgds', () => {
   test("should return { status: 404, msg: 'Not Found' }", (done) => {
